+++ android-sdk-couriers/src/main/java/com/citibox/courier/sdk/webview/models/WebAppEnvironment.kt
@@ -1,0 +1,5 @@
+package com.citibox.courier.sdk.webview.models
+
+enum class WebAppEnvironment {
+    Production, Sandbox, Test
+}