+++ app/src/main/java/com/citibox/courier/sdk/example/compose/MainScreen.kt
@@ -22,6 +22,7 @@ fun MainScreen(
         onPhoneChanged = viewModel::onPhoneChanged,
         onPhoneHashedChanged = viewModel::onPhoneHashedChanged,
         onDimensionsChanged = viewModel::onDimensionsChanged,
+        onEnvironmentChanged = viewModel::onEnvironmentChanged,
         onLaunchClicked = viewModel::onLaunch
     )